requirements/base.txt
#python==3.8.0
click==7.1.2
Flask==1.1.2
Flask-Script==2.0.6
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pony==0.7.13
Werkzeug==1.0.1
WTForms==2.1
flask-wtf-0.14.3
bootstrap-flask==1.5.1
